requirements.txt
streamlit
tensorflow==2.11
Pillow
numpy
plotly
matplotlib
